Fix for multiple display use while creating a VM
This should get you past the bug with multiple displays connected to the system when creating the VM per https://github.com/KhaosT/MacVM/issues/29. I've tested this change on a Mac Studio with two displays connected under 12.4. With the original code, I could not create a VM. With the change, I've created ~5 this afternoon.